+++ spec/fixtures/event.rb
